Increase est coverage
