+++ test/containers/Blog/Blog.test.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import Waypoint from 'react-waypoint';
-import { mount } from 'enzyme';
+import { mount, shallow } from 'enzyme';
 import Blog from '@containers/Blog/Blog';
 import Hero from '@components/Hero';
+import WithData from '@containers/Blog/index/'
 
 interface Post {
   data: {
@@ -188,3 +189,11 @@ describe('<Blog />', () => {
     });
   });
 });
+
+describe('With data', () => {
+  it('renders without crashing', () => {
+    const component = shallow(
+      <WithData />
+    );
+  })
+})
